==> 06_Species_meta.Rmd <==
---
title: "Pooled results approach_AOR_meta"
author: "Stefano Romano"
output: html_document
editor_options: 
  chunk_output_type: console
---
# Before you begin:

These scripts were tailored for the analyses we conducted. They are not meant to be used in other contexts and are reported here as a detailed extension of the method section of our article. Please, use them at your own risk.

If you use any of the material or approaches in these scripts, please, cite:
Romano et al, 2020, *Meta-analysis of the Parkinson’s disease gut microbiome suggests alterations linked to intestinal inflammation*

Thank you!


```{r setup, include=FALSE}

list.of.packages <- c("dplyr", "reshape2", "dplyr", "genodds", "meta")
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

if (!requireNamespace("BiocManager", quietly = TRUE)){
  install.packages("BiocManager")}


list.of.bioc <- c("phyloseq", "zCompositions", "propr", "microbiome")
new.packages <- list.of.bioc[!(list.of.bioc %in% installed.packages()[,"Package"])]
if(length(new.packages)) BiocManager::install(new.packages)


library(phyloseq)
library(dplyr)
library(microbiome)
library(reshape2)
library(devtools)
library(zCompositions)
library(propr)
library(meta)
library(genodds)

```

# Overview of the analyses
We will calculate the shift in abundances between PD and controls for each taxon.
For species, we will estimate the effect size using Agresti's generilized odd ratios.
We will use linear models (lm) for Genus, Family, and 16S-based functional predictions.
The effect size estimates will then be pooled using a random-effect meta-analysis.
Please provide the full path to the files and R objects that need to be loaded.

# Load data and format them for AOR
```{R}
l_phylo<-readRDS("Phyloseq_list_all_studies.rds")[-11] # I do not need the combined dataset

# Remove Eukaryota
l_phylo<-lapply(l_phylo, function(x) prune_taxa(taxa_names(x)[!(grepl("Eukaryota", taxa_names(x), ignore.case = TRUE))], x))
# Check that there are not Euk left
if(length(unlist(lapply(l_phylo, function(x) taxa_names(x)[grep("Eukary*", taxa_names(x))]))) == 0){
  print("All Eukaryotes removed")
}else{stop("There are still Eukaryotes sequences in the files")}
nam<-unique(unlist(lapply(l_phylo, taxa_names)))


# Check Chloroplasts
for(i in 1:length(l_phylo)){
  if(sum(grep("Chloroplast", taxa_names(l_phylo[[i]]))) > 0){
    l_phylo[[i]]<-prune_taxa(taxa_names(l_phylo[[i]])[!(grepl("Chloropl*", taxa_names(l_phylo[[i]]), ignore.case = TRUE))], l_phylo[[i]])
  } else {}
}

if(length(unlist(lapply(l_phylo, function(x) taxa_names(x)[grep("Chloropl*", taxa_names(x))]))) == 0){
  print("All Chloroplast removed")
}else{stop("There are still Chloroplast sequences in the files")}

# Remove samples that potentially have 0 counts
l_phylo<-lapply(l_phylo, function(x) prune_samples(sample_sums(x) > 0, x))

# Transform variables into factors
for(i in 1:length(l_phylo)){
  x<-meta(l_phylo[[i]])
  x$status<-as.factor(x$statu)
  x$study<-as.factor(x$study)
  sample_data(l_phylo[[i]])<-sample_data(x)
}

set.seed(987)


l_otu<-l_phylo
for(i in 1: length(l_phylo)){
  if(all(colnames(as.data.frame(otu_table(l_phylo[[i]]))) == sample_names(l_phylo[[i]]))){
  l_otu[[i]]<-as.data.frame(t(otu_table(l_phylo[[i]])))
  l_otu[[i]]$status<- get_variable(l_phylo[[i]])$status
  l_otu[[i]][l_otu[[i]]==0]<-1
    }else{stop("The order of samples in the maps and prop. is not the same")}
}

# Create a list of CLR objects
library(CoDaSeq)
l_phylo_sign<-l_phylo
l_clr_all<-lapply(l_otu, function(x) as.data.frame(codaSeq.clr(x[,-1*ncol(x)])))
for(i in 1:length(l_clr_all)){
  if(all(sample_names(l_phylo[[i]]) == rownames(l_clr_all[[i]]))){
    l_clr_all[[i]]$status<-meta(l_phylo[[i]])$status
  }else{stop("ORder of samples is not the same")}
}

id<-unlist(lapply(l_clr_all, rownames))
st<-unlist(lapply(l_clr_all, function(x) x$status))
status<-data.frame(status = st, id = id)

# Select only taxa detected in at least 3 datasets
appendnames <- subset(as.data.frame(table(Reduce(c, lapply(l_clr_all, names)))), Freq>=3)$Var1
allnamesF <- setdiff(as.character(appendnames), c("id", "Row.names", "status"))

# AOR for single dataset + meta analysis
meta_clrs_s <- lapply(allnamesF, function(a){
  print(a)
  results <- lapply(l_clr_all, FUN = function(d) {
    if(a%in%names(d)) {
      print(levels(d$status))
      lo <- genodds(d[,a], d$status)
      c(lo$pooled_lnodds, lo$pooled_SElnodds)
      } 
    else NULL
  })
  results <- data.frame(do.call(rbind,results))
  names(results) <- c("estimate", "se")
  print(results)
  metagen(TE=estimate, seTE=se,data=results,method.tau = "ML",title = a, studlab = rownames(results))
})

# Combine data
random.meta.f.all <- do.call(rbind,lapply(meta_clrs_s, function(m) as.data.frame(m[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")])))
random.meta.f.all$statistic <- allnamesF

# Save results
saveRDS(random.meta.f.all, file = "Meta_species.rds")

```

# Save workspace
```{R}
save.image("Meta_clr_lm_species.RData")
```
